Brought @material-ui/styles and @material-ui/core in-line with each other, plus added withRoot HOC to apply base theming
