+++ api/app/server.js
@@ -9,7 +9,7 @@ app.options('*', cors());
 
 app.use(routes);
 
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 8080;
 
 app.listen(port, () => {
   console.log(`App listening on port ${port}!`);
