+++ web/package.json
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.0.0-alpha.6",
-    "@material-ui/styles": "^3.0.0-alpha.10",
+    "@material-ui/styles": "^4.0.0-alpha.6",
     "axios": "^0.18.0",
     "jest-dom": "^3.1.3",
     "node-sass": "^4.11.0",
